Move environment tests to the correct source file
